\name{ict.var}
\alias{ict.var}
\title{Get values for select variables}
\usage{
ict.var(input, variable)
}
\arguments{
\item{input}{Path to the input .ict file}

\item{variable}{Variable (or vector of variables)}
}
\description{
Get values for select variables
}
\value{
Values of variable in data frame
}
\examples{
ict.var(input, variable)
}
