% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICTconverter.R
\name{ict_convert}
\alias{ict_convert}
\title{Convert and combine NASA ICARTT (.ict) files into a single CSV or XLSX,
while saving one metadata file from the first ICT file.}
\usage{
ict_convert(
  input_folder,
  output_file,
  output_format = c("csv", "xlsx"),
  missing_values = c(-9999, -999999, -7777, -8888)
)
}
\arguments{
\item{input_folder}{Path to the folder containing ICT files.}

\item{output_file}{Path to save the combined CSV/XLSX file.}

\item{output_format}{File format to output: "csv" (default) or "xlsx".}

\item{missing_values}{Numeric codes representing missing values to replace with NA.}
}
\value{
A data frame containing the combined data (invisibly).
}
\description{
Convert and combine NASA ICARTT (.ict) files into a single CSV or XLSX,
while saving one metadata file from the first ICT file.
}
\examples{
data_converted_csv <- ict_convert("ict_data_folder", "csv_data")
data_converted_xlsx <- ict_convert("ict_data_folder", "xlsx_data", "xlsx")
}
